+++ src/WorkflowEngineServiceProvider.php
@@ -10,6 +10,7 @@ use Illuminate\Contracts\Events\Dispatcher;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Support\ServiceProvider;
 use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowAdvanceCommand;
+use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowCancelCommand;
 use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowInstancesCommand;
 use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowListCommand;
 use JustSteveKing\WorkflowEngine\Console\Commands\WorkflowPruneCommand;
@@ -68,6 +69,7 @@ final class WorkflowEngineServiceProvider extends ServiceProvider
                 WorkflowStartCommand::class,
                 WorkflowSignalCommand::class,
                 WorkflowAdvanceCommand::class,
+                WorkflowCancelCommand::class,
                 WorkflowRetryCommand::class,
                 WorkflowTickCommand::class,
                 WorkflowPruneCommand::class,